revert(web): remove all core /src changes, keep only test files
