--- apps/web/src/lib/ThemeContext.ts
@@ -1,10 +1,0 @@
-import { createContext } from 'react';
-
-type Theme = 'light' | 'dark';
-
-export interface ThemeContextValue {
-  theme: Theme;
-  toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<ThemeContextValue | null>(null);
