--- apps/web/src/lib/useTheme.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react';
-import { ThemeContext } from './ThemeContext';
-import type { ThemeContextValue } from './ThemeContext';
-
-export function useTheme(): ThemeContextValue {
-  const ctx = useContext(ThemeContext);
-  if (!ctx) throw new Error('useTheme must be used within ThemeProvider');
-  return ctx;
-}
