+++ apps/web/src/pages/CardPage.tsx
@@ -24,31 +24,17 @@ export default function CardPage() {
 
   useEffect(() => {
     if (!id) return;
-
-    let cancelled = false;
-
-    const fetchCard = async () => {
-      try {
-        const data = await apiFetch<PublicCard>(`/api/u/card/${id}`);
-        if (!cancelled) {
-          setCard(data);
-          setError(null);
-        }
-      } catch {
-        if (!cancelled) {
-          setCard(null);
-          setError('Card not found');
-        }
-      } finally {
-        if (!cancelled) setLoading(false);
-      }
-    };
-
-    fetchCard();
-
-    return () => {
-      cancelled = true;
-    };
+    setLoading(true);
+    apiFetch<PublicCard>(`/api/u/card/${id}`)
+      .then((data) => {
+        setCard(data);
+        setError(null);
+      })
+      .catch(() => {
+        setCard(null);
+        setError('Card not found');
+      })
+      .finally(() => setLoading(false));
   }, [id]);
 
   // Update document title
