+++ apps/web/src/pages/ProfilePage.tsx
@@ -18,36 +18,27 @@ export default function ProfilePage() {
   const [profile, setProfile] = useState<PublicProfile | null>(null);
   const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(true);
+  const [mounted, setMounted] = useState(false);
   const [copyMessage, setCopyMessage] = useState('');
   const [copyStatus, setCopyStatus] = useState<'success' | 'error'>('success');
 
   useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  useEffect(() => {
     if (!username) return;
-
-    let cancelled = false;
-
-    const fetchProfile = async () => {
-      try {
-        const data = await apiFetch<PublicProfile>(`/api/u/${username}?source=web`);
-        if (!cancelled) {
-          setProfile(data);
-          setError(null);
-        }
-      } catch {
-        if (!cancelled) {
-          setProfile(null);
-          setError('User not found');
-        }
-      } finally {
-        if (!cancelled) setLoading(false);
-      }
-    };
-
-    fetchProfile();
-
-    return () => {
-      cancelled = true;
-    };
+    setLoading(true);
+    apiFetch<PublicProfile>(`/api/u/${username}?source=web`)
+      .then((data) => {
+        setProfile(data);
+        setError(null);
+      })
+      .catch(() => {
+        setProfile(null);
+        setError('User not found');
+      })
+      .finally(() => setLoading(false));
   }, [username]);
 
   async function copyProfileUrl() {
@@ -116,7 +107,7 @@ export default function ProfilePage() {
         className="bg-gradient"
         style={{ '--accent': profile.accentColor || '#6366f1' } as React.CSSProperties}
       />
-      <main className="profile-container loaded">
+      <main className={`profile-container ${mounted ? 'loaded' : ''}`}>
         <div
           className="profile-card glass"
           style={{ '--accent': profile.accentColor } as React.CSSProperties}
